refactor: unified transport and ledger contexts
